♻️ Refactored folder structure
